addressing some issues concerning model model

+++ src/SoundModules/Delay.js
@@ -44,7 +44,7 @@ class Delay extends Module{
             let feedbackLevels = this.inputs.feedback.getValues(recursion);
             let timeLevels = this.inputs.time.getValues(recursion);
 
-            this.cachedValues = new Float64Array(inputValues.length);
+            this.cachedValues = new Float32Array(inputValues.length);
             
             inputValues.forEach((value,sampleNumber)=>{
                 this.cachedValues[sampleNumber] = 0;
@@ -68,7 +68,6 @@ class Delay extends Module{
                 
             });
 
-            // this.changed({ cachedValues: this.cachedValues });
             //return this.cachedValues;
         };
     }
